refactor(backend): 整理 Tauri 应用外壳模块
