+++ src-tauri/src/app/hooks/crash.rs
