+++ src-tauri/src/app/hooks/mod.rs
@@ -1,0 +1,4 @@
+//! 应用生命周期钩子。
+
+pub(super) mod crash;
+pub(super) mod portable;
