+++ src-tauri/src/app/tray.rs
